Removed URL from response to GET at db level.
I was returning url: '/data/dbname/collectionname/' in items
from a fetch on a db returned from manageDatabase. The url has
changed now to include the mode and it seems wrong to include a
likely incorrect url here. The user never needs the url. So I
have changed MongoStore to not return the url field. This
required some changes to the Database editor where the url had
been used formerly. Function should be the same if I got it
right. The url didn't add any information that was really needed.